LPS-170922 Moves AssetEntryAction to asset-publisher-api
